deploy_ai_search/pyproject.toml
[project]
name = "deploy-ai-search"
version = "0.1.0"
description = "Helper functions to deploy Azure AI Search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-mgmt-web>=7.3.1",
    "azure-search-documents==11.6.0b8",
    "azure-storage-blob>=12.24.0",
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
text-2-sql-core = { workspace = true }
